Build: Extract shared bundler logic into @wordpress/build-core
Move the bundler-agnostic WordPress externals/handle mapping helpers into a
new @wordpress/build-core package and consume it from the
dependency-extraction-webpack-plugin and wp-build (esbuild) integrations,
removing the duplicated implementations.
