+++ packages/build-core/README.md
@@ -1,0 +1,7 @@
+# @wordpress/build-core
+
+Shared build utilities for WordPress bundler integrations.
+
+This package contains bundler-agnostic helpers for mapping WordPress package
+imports, vendor imports, and generated dependency metadata. It is intended for
+use by bundler-specific adapters such as webpack, esbuild, and Vite.
