+++ packages/wp-build/package.json
@@ -36,6 +36,7 @@
 	},
 	"dependencies": {
 		"@emotion/babel-plugin": "11.13.5",
+		"@wordpress/build-core": "file:../build-core",
 		"@wordpress/style-runtime": "file:../style-runtime",
 		"autoprefixer": "10.4.21",
 		"browserslist-to-esbuild": "2.1.1",
